Extract dumpHttp in new module
